change partition size to integer MBs

+++ vast_pipeline/forms.py
@@ -34,7 +34,7 @@ class PipelineRunForm(forms.Form):
     source_aggregate_pair_metrics_min_abs_vs = forms.FloatField()
     num_workers = forms.IntegerField(required=False)
     num_workers_io = forms.IntegerField()
-    max_partition_mb = forms.FloatField()
+    max_partition_mb = forms.IntegerField()
     pair_metrics = forms.BooleanField(required=False)
     use_condon_errors = forms.BooleanField(required=False)
     create_measurements_arrow_files = forms.BooleanField(required=False)
